+++ src/rviz/default_plugin/map_display.cpp
@@ -701,7 +701,7 @@ void MapDisplay::showMap()
   }
 
   bool map_status_set = false;
-  if( width * height != current_map_.data.size() )
+  if( width * height != static_cast<int>(current_map_.data.size()) )
   {
     std::stringstream ss;
     ss << "Data size doesn't match width*height: width = " << width
@@ -710,7 +710,7 @@ void MapDisplay::showMap()
     map_status_set = true;
   }
 
-  for(int i=0;i<swatches.size();i++){
+  for(size_t i=0;i<swatches.size();i++){
     swatches[i]->updateData();
 
     Ogre::Pass* pass = swatches[i]->material_->getTechnique(0)->getPass(0);
